+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Fixes:
 Enterprise Fixes:
 - [journey_engine] Avoid throwing on duplicate events and prevent overwriting existing event map entries during event creation
 
+Enterprise Features:
+- [block] Allow using regex for event filter
+
 ## Version 25.03.47
 Fixes:
 - [content] Bugfixes for content showing